@@ -1,8 +1,7 @@
 cmake_minimum_required(VERSION 3.16)
 
-# uncomment this if you wand debug binary
 SET( CMAKE_BUILD_TYPE Release)
-set (CMAKE_CXX_STANDARD 14)
+set (CMAKE_CXX_STANDARD 17)
 
 project(n-puzzle)
 
