@@ -1,13 +1,16 @@
 #include "Board.class.hpp"
 #include <iostream>
 #include <cmath>
+#include <algorithm>
+
+
 /*
 ** a printer helper function
 */
 
 void Board::print()
 {
-    for (auto i : body)
+    for (auto i : state)
         std::cout << i << " ";
     std::cout << "\n";
 }
@@ -23,7 +26,7 @@ void Board::print()
 
 Board Board::gen_solution(int size)
 {
-    std::vector<int> body(size * size);
+    std::vector<int> state(size * size);
     int inc = 1;
     int li = 0, ri = size - 1, ui = 0, di = size - 1;
     /*
@@ -38,34 +41,34 @@ Board Board::gen_solution(int size)
         if (dir % 4 == 0)
         {
             for (int i = li; i <= ri; i++)
-                body[i + size * ui] = inc++;
+                state[i + size * ui] = inc++;
             ui++;
         }
         // right columns travers
         else if (dir % 4 == 1)
         {
             for (int i = ui; i <= di; i++)
-                body[ri + size * i] = inc++;
+                state[ri + size * i] = inc++;
             ri--;
         }
         // down rows travers
         else if (dir % 4 == 2)
         {
             for (int i = ri; i >= li; i--)
-                body[i + size * di] = inc++;
+                state[i + size * di] = inc++;
             di--;
         }
         // left columns travers
         else if (dir % 4 == 3)
         {
             for (int i = di; i >= ui; i--)
-                body[li + size * i] = inc++;
+                state[li + size * i] = inc++;
             li++;
         }
         dir++;
     }
 
-    return Board(body, size);
+    return Board(state, size);
 }
 
 /**
@@ -73,28 +76,28 @@ Board Board::gen_solution(int size)
  * and return the boolean result
  * */
 
-bool    Board::is_solvable()
+bool Board::is_solvable()
 {
-    int                 index;
-    int                 time;
-    int                 diff;
-    std::pair<int,int>  zero_in_sol;
-    std::pair<int,int>  zero_in_board;
-    Board               solution = gen_solution(size);
+    int index;
+    int time;
+    int diff;
+    std::pair<int, int> zero_in_sol;
+    std::pair<int, int> zero_in_board;
+    Board solution = gen_solution(size);
 
     zero_in_sol = solution.get_position_of_zero();
     zero_in_board = get_position_of_zero();
     time = 0;
     diff = abs(zero_in_sol.first - zero_in_board.first) + abs(zero_in_sol.second - zero_in_board.second);
-    for (int i = 0; i< solution.body.size(); i++)
+    for (int i = 0; i < solution.state.size(); i++)
     {
-        for (int j = i; j < body.size(); j++)
+        for (int j = i; j < state.size(); j++)
         {
-            if (body[j] == solution.body[i])
+            if (state[j] == solution.state[i])
             {
                 if (j != i)
                 {
-                    std::swap(body[j], body[i]);
+                    std::swap(state[j], state[i]);
                     time++;
                 }
             }
@@ -111,17 +114,24 @@ bool    Board::is_solvable()
 
 std::pair<int, int> Board::get_position_of_zero()
 {
-    int                 index;
+    int index;
     std::pair<int, int> ans;
 
     index = 0;
-    for (int i=0; i < body.size(); i++)
-        if (body[i] == 0)
+    for (int i = 0; i < state.size(); i++)
+        if (state[i] == 0)
         {
-            index = i+1;
+            index = i + 1;
             break;
         }
     ans.first = (index % size == 0) ? size : index % size;
     ans.second = (index - ans.first) / size;
     return ans;
+}
+
+std::vector<int> Board::create_new(int index1, int index2)
+{
+    std::vector<int> newState(state);
+    std::swap(newState[index1], newState[index2]);
+    return newState;
 }
